refactor: refactor structure
